Break getDetails function to smaller functions
